apps/dash_dbsql/requirements.txt
dash
databricks-sdk==0.33.0
pandas==2.2.3
